# docs/research_openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Algo Trading Platform",
  "description": "Algo trading platform for Indian markets. Default mode is PAPER; live trading is disabled unless explicitly enabled.",
  "version": "0.1.0"
 },
 "paths": {
  "/api/v1/research/run": {
   "post": {
    "tags": [
     "Research"
    ],
    "summary": "Start Run",
    "description": "Start a research scan in the background. Poll GET /research/runs/{id} for progress.",
    "operationId": "start_run_api_v1_research_run_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_ResearchRunRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/research/summary": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Summary",
    "operationId": "summary_api_v1_research_summary_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_ResearchSummaryRead_"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/research/runs": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "List Runs",
    "operationId": "list_runs_api_v1_research_runs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 30,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_list_ResearchRunRead__"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/runs/{run_id}": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Get Run",
    "operationId": "get_run_api_v1_research_runs__run_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_ResearchRunRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/candidates": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Candidates",
    "description": "Top Research Candidates from the latest completed run (or `run_id`).",
    "operationId": "candidates_api_v1_research_candidates_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     },
     {
      "name": "min_score",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "maximum": 100,
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Score"
      }
     },
     {
      "name": "min_confidence",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "maximum": 100,
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Confidence"
      }
     },
     {
      "name": "min_rel_volume",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Rel Volume"
      }
     },
     {
      "name": "sector",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sector"
      }
     },
     {
      "name": "risk",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Risk"
      }
     },
     {
      "name": "direction",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "BULLISH",
          "BEARISH",
          "NEUTRAL"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Direction"
      }
     },
     {
      "name": "min_price",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Price"
      }
     },
     {
      "name": "max_price",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Max Price"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "score",
        "confidence",
        "rel_volume",
        "change",
        "atr",
        "price",
        "symbol",
        "initial"
       ],
       "type": "string",
       "default": "score",
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "desc",
        "asc"
       ],
       "type": "string",
       "default": "desc",
       "title": "Order"
      }
     },
     {
      "name": "include_unanalyzed",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Unanalyzed"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 200,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_CandidateList_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/candidates/{symbol}": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Candidate",
    "operationId": "candidate_api_v1_research_candidates__symbol__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "symbol",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Symbol"
      }
     },
     {
      "name": "run_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_CandidateRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/market": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Market",
    "operationId": "market_api_v1_research_market_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_MarketOverview_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/sectors": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Sectors",
    "operationId": "sectors_api_v1_research_sectors_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_SectorOverview_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/weights": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "List Weights",
    "operationId": "list_weights_api_v1_research_weights_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_list_WeightSetRead__"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Research"
    ],
    "summary": "Propose Weights",
    "description": "Propose a new weight set. It changes nothing until an administrator approves it.",
    "operationId": "propose_weights_api_v1_research_weights_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WeightSetCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_WeightSetRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/research/weights/{weight_id}/approve": {
   "post": {
    "tags": [
     "Research"
    ],
    "summary": "Approve Weights",
    "operationId": "approve_weights_api_v1_research_weights__weight_id__approve_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "weight_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Weight Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_WeightSetRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/source-registry": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Source Registry",
    "operationId": "source_registry_api_v1_research_source_registry_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_list_SourceRegistryRead__"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/research/source-registry/{source_id}": {
   "put": {
    "tags": [
     "Research"
    ],
    "summary": "Update Source",
    "operationId": "update_source_api_v1_research_source_registry__source_id__put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "source_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Source Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SourceRegistryUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_SourceRegistryRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/universes": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Universes",
    "operationId": "universes_api_v1_research_universes_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_list_UniverseRead__"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/research/universes/{name}": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Universe",
    "operationId": "universe_api_v1_research_universes__name__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "name",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Name"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_UniverseRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "Research"
    ],
    "summary": "Replace Universe",
    "operationId": "replace_universe_api_v1_research_universes__name__put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "name",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Name"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UniverseWrite"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_UniverseRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/runs/{run_id}/calibrate": {
   "post": {
    "tags": [
     "Research"
    ],
    "summary": "Calibrate Run",
    "description": "Measure the market that followed each score in this run. Needs an hour of trading to have passed.",
    "operationId": "calibrate_run_api_v1_research_runs__run_id__calibrate_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "force",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Force"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_dict_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/runs/{run_id}/calibration": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Run Calibration",
    "operationId": "run_calibration_api_v1_research_runs__run_id__calibration_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_dict_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/research/calibration": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Calibration",
    "description": "Outcomes by score bucket across every calibrated run. Measured results, not a prediction.",
    "operationId": "calibration_api_v1_research_calibration_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_dict_"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/research/calibration/pending": {
   "get": {
    "tags": [
     "Research"
    ],
    "summary": "Calibration Pending",
    "operationId": "calibration_pending_api_v1_research_calibration_pending_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiResponse_list_dict__"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ApiResponse_ResearchRunRead_": {
    "properties": {
     "success": {
      "type": "boolean",
      "title": "Success",
      "default": true
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "status": {
      "$ref": "#/components/schemas/StatusInfo"
     },
     "data": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ResearchRunRead"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "message",
     "status"
    ],
    "title": "ApiResponse[ResearchRunRead]"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RunRequest": {
    "properties": {
     "market": {
      "type": "string",
      "const": "NSE",
      "title": "Market",
      "default": "NSE"
     },
     "universe": {
      "type": "string",
      "maxLength": 30,
      "minLength": 1,
      "title": "Universe",
      "default": "NIFTY50"
     },
     "symbols": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Symbols",
      "description": "Required when universe is CUSTOM"
     },
     "research_type": {
      "type": "string",
      "const": "INTRADAY",
      "title": "Research Type",
      "default": "INTRADAY"
     },
     "depth": {
      "type": "string",
      "enum": [
       "QUICK",
       "STANDARD",
       "DEEP"
      ],
      "title": "Depth",
      "default": "STANDARD"
     },
     "as_of": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of",
      "description": "Analyse the market as it was at this time. Defaults to now."
     }
    },
    "type": "object",
    "title": "RunRequest"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "JWT from POST /api/v1/auth/login",
    "scheme": "bearer"
   }
  }
 }
}